Unified apps structure, updated request templates
