--- buddy_mentorship/templates/buddy_mentorship/request.html
@@ -1,14 +1,0 @@
-{% extends 'buddy_mentorship/base.html' %}
-{% block content %}
-<h2>
-    Request
-</h2>
-
-<p>
-This is a request from {{ buddy_request.requestor }} to {{ buddy_request.requestee }}
-</p>
-
-<p>
-{{ buddy_request.message }}
-</p>
-{% endblock %}
